+++ unstructured_ingest/v2/processes/connectors/assets/databricks_delta_table_schema.sql
@@ -1,7 +1,7 @@
 CREATE TABLE elements (
     id STRING NOT NULL PRIMARY KEY,
-    record_id STRING,
-    element_id STRING,
+    record_id STRING NOT NULL,
+    element_id STRING NOT NULL,
     text STRING,
     embeddings ARRAY<FLOAT>,
     type STRING,